bot/requirements.txt
aiohttp==3.9.1
aiosqlite==0.19.0
rich==13.7.0
python-dotenv==1.0.1
